requirements.txt
python-telegram-bot==20.7
Pillow==10.2.0
piexif==1.1.3
Flask==3.0.2
requests==2.31.0
python-dotenv==1.0.0
